Move models to keys folder
