+++ src/Auth0.ManagementApi/Clients/KeysClient.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+using Auth0.ManagementApi.Models.Keys;
 
 namespace Auth0.ManagementApi.Clients
 {
