+++ src/Auth0.ManagementApi/Models/Keys/RotateSigningKeyResponse.cs
@@ -1,17 +1,17 @@
 using Newtonsoft.Json;
 
-namespace Auth0.ManagementApi.Models
+namespace Auth0.ManagementApi.Models.Keys
 {
-    public class RevokeSigningKeyResponse
+    public class RotateSigningKeyResponse
     {
         /// <summary>
-        /// The id of the revoked signing key
+        /// The id of the next signing key
         /// </summary>
         [JsonProperty("kid")]
         public string Kid { get; set; }
 
         /// <summary>
-        /// The public certificate of the revoked signing key
+        /// The public certificate of the next signing key
         /// </summary>
         [JsonProperty("cert")]
         public string Cert { get; set; }
